optional wrapper for shared more ops
