+++ packages/ui-shared-lite/MoreOptions/style.scss
@@ -1,3 +1,7 @@
+.moreOptionsWrapper {
+  position: relative;
+}
+
 .moreBtn {
   background: none;
   padding: 0;
