+++ packages/ui-shared-lite/MoreOptions/style.scss.d.ts
@@ -7,6 +7,7 @@ interface CssExports {
   'moreBtnDots': string;
   'moreMenu': string;
   'moreMenuItem': string;
+  'moreOptionsWrapper': string;
   'noHover': string;
 }
 declare var cssExports: CssExports;
